requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
coreapi==2.0.0
coverage==4.2
decorator==4.0.10
dj-database-url==0.4.1
Django==1.9
django-braces==1.9.0
django-envie==0.0.9
django-nose==1.4.4
django-oauth-toolkit==0.10.0
django-rest-framework-social-oauth2==1.0.4
django-rest-swagger==0.3.5
djangorestframework==3.4.6
enum34==1.1.6
gunicorn==19.6.0
ipython==5.1.0
ipython-genutils==0.1.0
itypes==1.1.0
nose==1.3.7
oauthlib==1.0.3
openapi-codec==1.0.0
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.3.1
postgres==2.2.1
prompt-toolkit==1.0.7
psycopg2==2.6.1
ptyprocess==0.5.1
Pygments==2.1.3
PyJWT==1.4.2
python-openid==2.2.5
python-social-auth==0.2.21
PyYAML==3.12
requests==2.11.1
requests-oauthlib==0.6.2
rest-social-auth==0.5.0
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
traitlets==4.3.0
uritemplate==3.0.0
wcwidth==0.1.7
whitenoise==3.2.1
